requirements.txt
asn1crypto==0.24.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
cryptography==2.1.4
DAWG-Python==0.7.2
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.404381.4453942
Flask==0.12.2
Flask-Mail==0.9.1
textblob==0.15.1
python-docx==0.8.6
passlib==1.7.1
requests==2.18.4
WTForms==2.1
mysqlclient==1.3.10
Jinja2==2.10
urllib3==1.22
gunicorn==19.7.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
yadisk==1.2.4
YaDiskClient==0.5.0
